refact : 유저가 작성한 게시글 조회 로직 이동
profiles -> posts
