+++ repo/profiles/schemas.py
@@ -1,5 +1,4 @@
 from drf_spectacular.utils import (
-    OpenApiParameter,
     OpenApiResponse,
     extend_schema,
     extend_schema_view,
@@ -9,8 +8,6 @@ from repo.profiles.serializers import (
     UserProfileSerializer,
     UserUpdateSerializer,
 )
-from repo.records.models import Post
-from repo.records.posts.serializers import UserPostSerializer
 from repo.records.serializers import UserNoteSerializer
 
 Profile_Tag = "profiles"
@@ -74,29 +71,6 @@ class OtherProfileSchema:
     )
 
     other_proflie_schema_view = extend_schema_view(get=other_proflie_get_schema)
-
-
-class UserPostListSchema:
-    user_post_list_get_schema = extend_schema(
-        parameters=[
-            OpenApiParameter(
-                name="subject",
-                type=str,
-                enum=[choice[0] for choice in Post.SUBJECT_TYPE_CHOICES],
-                description="게시글 주제",
-            ),
-        ],
-        summary="유저 게시글 조회",
-        description="특정 사용자의 게시글을 주제별로 조회합니다. (정렬 기준: 최신순)",
-        responses={
-            200: UserPostSerializer(many=True),
-            400: OpenApiResponse(description="Invalid subject parameter"),
-            404: OpenApiResponse(description="Not Found"),
-        },
-        tags=[Profile_Records_Tag],
-    )
-
-    user_post_list_schema_view = extend_schema_view(get=user_post_list_get_schema)
 
 
 class UserNoteSchema:
