+++ repo/profiles/urls.py
@@ -16,6 +16,5 @@ urlpatterns = [
     path("<int:id>/", views.OtherProfileAPIView.as_view(), name="other_profile"),
     path("api/token/", TokenObtainPairView.as_view(), name="token_obtain_pair"),
     path("api/token/refresh/", TokenRefreshView.as_view(), name="token_refresh"),
-    path("<int:id>/posts/", views.UserPostListAPIView.as_view(), name="user_posts"),
     path("<int:id>/notes/", views.UserNoteAPIView.as_view(), name="user_notes"),
 ]
